+++ include/ed/loop_usage_status.h
@@ -122,10 +122,10 @@ public:
         boost::mutex::scoped_lock const lock(lock_);
         long double const curtime = tue::Timer::nowMicroSec(); // Micro-seconds
         long double const total_loop_time = timer_.getTotalLoopTime(); // Seconds
-        int const curseq = timer_.getIterationCount();
+        int const curseq = static_cast<int>(timer_.getIterationCount());
         int const events = curseq - seq_nums_[hist_indx_];
-        double const window = (curtime - starts_[hist_indx_]) * 0.000001; // Micro-seconds -> Seconds
-        double const loop_time_window = total_loop_time - durations_[hist_indx_]; // Seconds
+        double const window = static_cast<double>((curtime - starts_[hist_indx_]) * 0.000001); // us -> s
+        double const loop_time_window = static_cast<double>(total_loop_time - durations_[hist_indx_]); // Seconds
         double freq = 0;
 
         if (window != 0)
