+++ include/ed/time.h
@@ -18,12 +18,12 @@ public:
 
     friend std::ostream& operator<<(std::ostream& out, const Time& d)
     {
-        int const isecs = d.secs_;
+        int const isecs = static_cast<int>(d.secs_);
 
         int const h = isecs / 3600;
         int const m = (isecs / 60) % 60;
         int const s = isecs % 60;
-        int const ms = 1000 * (d.secs_ - isecs);
+        int const ms = static_cast<int>(1000 * (d.secs_ - isecs));
 
         out << h << ":" << m << ":" << s << ":" << ms;
